pyproject.toml
[tool.poetry]
name = "procman"
version = "0.1.0"
description = ""
authors = ["Atsuya Ide <atsuya.ide528@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
schedule = "^1.1.0"
watchdog = "^2.1.6"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
